Change nonsense uint to int

+++ LiquidCrystalNationalCharsets.cpp
@@ -8,7 +8,7 @@ void createChar(LiquidCrystal lcd, uint8_t location, uint8_t charmap[], uint8_t 
   insteadOff[location]=instead;
 }
 
-void writeLCD(LiquidCrystal lcd, uint8_t outChar, uint8_t column, uint8_t row) {
+void writeLCD(LiquidCrystal lcd, uint8_t outChar, int column, int row) {
   
   if ( column>=0 && row>=0 ) {
     lcd.setCursor(column, row);
@@ -17,16 +17,18 @@ void writeLCD(LiquidCrystal lcd, uint8_t outChar, uint8_t column, uint8_t row) {
   for (uint8_t i=0; i<8; i++) {
     if (outChar==insteadOff[i]) outChar=uint8_t(i);
   }
+
   lcd.write(outChar);
 }
 
-void writeLCD(LiquidCrystal lcd, String text, uint8_t column, uint8_t row) {
+void writeLCD(LiquidCrystal lcd, String text, int column, int row) {
   
-  if ( column>=0 && row>=0 ) {
+  if ( column >= 0 && row >= 0 ) {
     lcd.setCursor(column, row);
   }
   
   for(int i=0; i<text.length(); i++) {
-    writeLCD( lcd, text.charAt(i),-1, -1);
+    writeLCD( lcd, text.charAt(i), -1, -1);
+    //writeLCD( lcd, text.charAt(i), i, 0);
   }
 }
